LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/semantics/popovers/popover-anchor-transition.tentative-expected.txt


FAIL Transitioning display property of an anchored popover should not cause a position shift assert_equals: Should not shift in x axis expected 367 but got 0
